+++ src/index.ts
@@ -7,8 +7,8 @@ export {ComponentRouteAuthorized} from './common/componentRouteAuthorized.decora
 export {AuthorizeDirective} from './common/authorize.directive';
 export {AccessToken} from './common/accessToken';
 export {LetAuthorizedContext, LetAuthorizedDirective} from './common/letAuthorized.directive';
-export {AuthorizationModule} from './modules/authorization.module';
 export {AuthInterceptorOptions} from './authInterceptor/authInterceptor.options';
 export {evaluatePermissions, isAuthorized} from './misc/utils';
 export * from './authInterceptor/authInterceptor';
 export * from './suppressAuthInterceptor/suppressAuthInterceptor';
+export * from './pipes';